Move relational.js back to knex.js, and put them in the same directory
